Code Restructuring & Clean-up
