+++ package.json
@@ -43,6 +43,7 @@
     "@web3-react/portis-connector": "^6.0.9",
     "@web3-react/walletconnect-connector": "^6.2.0",
     "@web3-react/walletlink-connector": "^6.0.9",
+    "@fuseio/rewards-sdk": "^0.1.8",
     "ajv": "^6.12.3",
     "bignumber.js": "^9.0.1",
     "cids": "^1.0.0",
