+++ src/components/FarmTable/icons.tsx
@@ -1,11 +1,11 @@
+import styled from 'styled-components'
+import React from 'react'
 import wethfuse from '../../assets/svg/pairs/WETH-FUSE.svg'
 import goodusdc from '../../assets/svg/pairs/G$-USDC.svg'
 import fusdbnb from '../../assets/svg/pairs/FUSD-BNB.svg'
 import fusdfuse from '../../assets/svg/pairs/FUSD-FUSE.svg'
-import styled from 'styled-components'
-import React from 'react'
 
-export default function Icon(props: { name: string; contract: string }) {
+export default function Icon(props: { name: string; address: string }) {
   const Map: { [name: string]: any } = {
     '0x04Ee5DE43332aF99eeC2D40de19962AA1cC583EC': goodusdc,
     '0x65995B106988E9aCd15998a5DF95aDe89b6511c8': fusdbnb,
@@ -33,7 +33,7 @@ export default function Icon(props: { name: string; contract: string }) {
 
   return (
     <Container>
-      <img src={Map[props.contract]} alt="" />
+      <img src={Map[props.address]} alt="" />
       <span>{props.name}</span>
     </Container>
   )
