+++ src/components/RewardCards/claim.tsx
@@ -6,15 +6,26 @@ import { ButtonPrimary } from '../Button'
 import { useTransactionAdder } from '../../state/transactions/hooks'
 
 const Container = styled('div')`
+  display: flex;
+  flex-wrap: wrap;
   text-align: left;
   margin-top: 10px;
   font-size: 1.25rem;
   padding: 1.5rem;
   padding-top: 0rem;
   padding-bottom: 0rem;
-  display: flex;
-  flex-wrap: wrap;
 `
+
+const Wrapper = styled('div')`
+  display: flex;
+  flex: wrap;
+  margin: auto;
+  margin-bottom: 15px;
+  margin-top: 15px
+  overflow: hidden;
+  text-align: center;
+`
+
 const Pool = styled('div')`
   padding: 0.25rem;
   width: 100%;
@@ -30,6 +41,12 @@ const Pool = styled('div')`
     text-align: left;
   }
 `
+const Claim = styled('div')`
+  margin-top: 24px;
+  margin-bottom: 24px;
+  width: 100%;
+  font-size: 14px;
+`
 
 const Button = styled(ButtonPrimary)`
   border-radius: 12px;
@@ -38,29 +55,6 @@ const Button = styled(ButtonPrimary)`
   }
 `
 
-const Claims = styled('div')`
-  margin-top: 24px;
-  margin-bottom: 24px;
-  width: 100%;
-  font-size: 14px;
-`
-const Collected = styled('div')`
-  margin-top: 24px;
-  margin-bottom: 24px;
-  width: 100%;
-  font-size: 14px;
-  text-align: center;
-`
-
-const Wrapper = styled('div')`
-display: flex;
-flex: wrap;
-margin: auto;
-margin-bottom: 15px;
-margin-top: 15px
-    overflow: hidden;
-    text-align: center;
-`
 interface ClaimProps {
   withdrawValue?: number
   data: {
@@ -79,11 +73,8 @@ interface ClaimProps {
 }
 
 export default function ClaimReward(props: ClaimProps) {
-  const [contract, setContract] = useState<{ stakingContractAddress: string; tokenAddress: string; user: string }>({
-    stakingContractAddress: '',
-    tokenAddress: '',
-    user: ''
-  })
+  const addTransaction = useTransactionAdder()
+  const { account, library } = useActiveWeb3React()
   const [result, setResult] = useState<{
     rewardAcruded: string
     rewardUnlockedUser: string
@@ -91,8 +82,6 @@ export default function ClaimReward(props: ClaimProps) {
     rewardTotal: string
     rewardUnlocked: string
   }>({ rewardAcruded: '0', rewardUnlockedUser: '0', rewardEstimate: '0', rewardTotal: '0', rewardUnlocked: '0' })
-  const addTransaction = useTransactionAdder()
-  const { account, library } = useActiveWeb3React()
 
   async function onClaim(contract: any) {
     if (!library || !account) return
@@ -114,13 +103,12 @@ export default function ClaimReward(props: ClaimProps) {
   }
 
   useEffect(() => {
-    setContract(props.contract)
     setResult(props.data)
   }, [props, addTransaction])
 
   return (
-    <div>
-      <Container>
+    <Container>
+      <Wrapper>
         <Pool>
           <b>Total Rewards</b> <span>{Number(result.rewardTotal) || 0}</span>
         </Pool>
@@ -130,20 +118,20 @@ export default function ClaimReward(props: ClaimProps) {
         <Pool>
           <b>Your Estimated Rewards</b> <span>{Number(result.rewardEstimate).toFixed(0) || 0}</span>
         </Pool>
+      </Wrapper>
 
-        <Wrapper>
-          <Claims>
-            <b>Claimed</b> {result.rewardAcruded || 0}
-          </Claims>
-          <Claims>
-            <b>Claimable</b> {result.rewardUnlockedUser || 0}
-          </Claims>
-          <Collected>
-            <b>Collected</b> {(Number(result.rewardAcruded) || 0 + Number(result.rewardUnlockedUser)).toFixed(2) || 0}
-          </Collected>
-        </Wrapper>
-      </Container>
-      <Button onClick={() => onClaim(contract)}> Claim WFUSE Rewards</Button>
-    </div>
+      <Wrapper>
+        <Claim>
+          <b>Claimed</b> {result.rewardAcruded || 0}
+        </Claim>
+        <Claim>
+          <b>Claimable</b> {result.rewardUnlockedUser || 0}
+        </Claim>
+        <Claim style={{ textAlign: 'center' }}>
+          <b>Collected</b> {(Number(result.rewardAcruded) || 0 + Number(result.rewardUnlockedUser)).toFixed(2) || 0}
+        </Claim>
+      </Wrapper>
+      <Button onClick={() => onClaim(props.contract)}> Claim WFUSE Rewards</Button>
+    </Container>
   )
 }
