+++ src/components/RewardCards/index.tsx
@@ -1,9 +1,10 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import styled from 'styled-components'
 import { useActiveWeb3React } from '../../hooks'
 import Claim from './claim'
 import Withdraw from './withdraw'
 import Deposit from './deposit'
+import { Flex } from 'rebass'
 
 const Wrapper = styled('div')`
   width: 100%;
@@ -20,8 +21,13 @@ const Wrapper = styled('div')`
   border-radius: 12px;
 `
 
-const Item = styled('div')`
-  padding: 0.25rem;
+const TabGroup = styled.div`
+  display: flex;
+  border-radius: 16px;
+  width: 100%;
+  height: 48px;
+  margin-bottom: 14px;
+  background: #111219;
 `
 
 const Tab = styled.button<{ active: any }>`
@@ -36,34 +42,23 @@ const Tab = styled.button<{ active: any }>`
   border-radius: 0;
   ${({ active }) =>
     active &&
-    `
-    :before{
+    `:before{
       content:"";
-  position:absolute;
-  width: 95%;
-  left:0;
-  bottom:0;
-  top:0;
-  border-radius:16px; 
-  padding:3px; 
-  background:linear-gradient(110deg, #b1ffbf 7%, #fff16d);
-  -webkit-mask: 
-     linear-gradient(#fff 0 0) content-box, 
-     linear-gradient(#fff 0 0);
-  -webkit-mask-composite: destination-out; 
-  mask-composite: exclude; 
-    }
-
+      position:absolute;
+      width: 95%;
+      left:0;
+      bottom:0;
+      top:0;
+      border-radius:16px; 
+      padding:3px; 
+      background:linear-gradient(110deg, #b1ffbf 7%, #fff16d);
+      -webkit-mask: 
+        linear-gradient(#fff 0 0) content-box, 
+        linear-gradient(#fff 0 0);
+      -webkit-mask-composite: destination-out; 
+      mask-composite: exclude; 
+        }
   `}
-`
-
-const ButtonGroup = styled.div`
-  display: flex;
-  border-radius: 16px;
-  width: 100%;
-  height: 48px;
-  margin-bottom: 14px;
-  background: #111219;
 `
 
 const types = ['Deposit', 'Withdraw', 'Stats']
@@ -93,47 +88,23 @@ function RewardsReselect(param: string, result: any, contract: any) {
 }
 
 export default (props: RewardSelection) => {
-  const [contract, setContract] = useState<{
-    stakingContractAddress: string
-    tokenAddress: string
-    user: string
-    token0: string
-    token1: string
-  }>({ stakingContractAddress: '', tokenAddress: '', user: '', token0: '', token1: '' })
-  const [result, setResult] = useState<any>({
-    res: [],
-    lpTotal: '0',
-    rewardsTotal: '0',
-    lpUser: '0',
-    rewardsUnlocked: '0',
-    rewardAcruded: '0',
-    rewardUnlockedUser: '0',
-    rewardEstimate: '0',
-    rewardTotal: '0',
-    rewardUnlocked: '0'
-  })
   const [active, setActive] = useState(types[0])
   const { chainId } = useActiveWeb3React()
-
-  useEffect(() => {
-    setContract(props.contract)
-    setResult(props.result)
-  }, [props])
 
   if (chainId === 122) {
     return (
       <Wrapper>
-        <ButtonGroup>
+        <TabGroup>
           {types.map(type => (
             <Tab key={type} active={active === type} onClick={() => setActive(type)}>
               {type}
             </Tab>
           ))}
-        </ButtonGroup>
-        {RewardsReselect(active, result, contract)}
+        </TabGroup>
+        {RewardsReselect(active, props.result, props.contract)}
       </Wrapper>
     )
   } else {
-    return <Item>Please, switch to Fuse Network to interact with LP Farm</Item>
+    return <Flex padding={'15px'}>Please, switch to Fuse Network to interact with LP Farm</Flex>
   }
 }
