+++ src/components/RewardCards/info.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components'
 import { ButtonPrimary } from '../Button'
 import Modal from '../Modal'
 import Questionmark from '../../assets/svg/questionmark.svg'
+import { Content, Header, HeaderText, StyledModal } from './modal'
 
 const Container = styled('div')<{ color: string; txt: string }>`
   display: flex;
@@ -77,52 +78,8 @@ const Title = styled('div')<{ txt: string }>`
   line-height: 18px;
   color: ${({ txt }) => txt};
 `
-export const Backdrop = styled.div`
-  position: fixed;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.3);
-  z-index: 500;
-`
-export const StyledModal = styled.div`
-  z-index: 100;
-  padding: 24px;
-  background: #242637;
-  position: relative;
-  margin: auto;
-  border-radius: 12px;
-`
-export const Header = styled.div`
-  border-radius: 8px 8px 0 0;
-  display: flex;
-  justify-content: space-between;
-`
 
-export const HeaderText = styled.div`
-  color: #fff;
-  align-self: center;
-  color: lightgray;
-`
-
-export const Content = styled.div`
-  padding-bottom: 15px;
-  max-height: 30rem;
-  overflow-x: hidden;
-  overflow-y: auto;
-  > h1 {
-    font-size: 24px;
-    font-weight: 600;
-  }
-  > p {
-    font-size: 14px;
-    font-weight: 400;
-    line-height: 21px;
-  }
-`
-
-const Item2 = styled('div')`
+const TitleModal = styled(HeaderText)`
   display: flex;
   width: 100%;
   text-align: center;
@@ -130,7 +87,7 @@ const Item2 = styled('div')`
   position: relative;
 `
 
-const IconW = styled('div')`
+const IconModal = styled('div')`
   border-radius: 999px;
   background-color: #ffffff1a;
   height: 48px;
@@ -179,17 +136,15 @@ export default function RewardsAPY(props: any) {
         <Wrapper aria-modal aria-labelledby={'APY'} tabIndex={-1} role="dialog">
           <StyledModal>
             <Header>
-              <HeaderText>
-                <Item2
-                  onClick={() => {
-                    setOpen(false)
-                  }}
-                >
-                  <IconW>
-                    <img src={Questionmark} width="28px" height="28px" alt="APY Icon"></img>
-                  </IconW>
-                </Item2>
-              </HeaderText>
+              <TitleModal
+                onClick={() => {
+                  setOpen(false)
+                }}
+              >
+                <IconModal>
+                  <img src={Questionmark} width="28px" height="28px" alt="APY Icon"></img>
+                </IconModal>
+              </TitleModal>
             </Header>
             <Content>
               <h1>What does &quot; APY &quot; mean?</h1>
