+++ src/pages/App.tsx
@@ -13,7 +13,7 @@ import {
 } from './AddLiquidity/redirects'
 import Pool from './Pool'
 import Farm from './Farm'
-import FarmSelect from './Farm/reselect'
+import FarmReward from './Farm/reward'
 import Lending from './Lending'
 import Home from './Home'
 import PoolFinder from './PoolFinder'
@@ -90,7 +90,7 @@ export default function App() {
                   <Route exact strict path="/remove/:currencyIdA/:currencyIdB" component={RemoveLiquidity} />
                   <Route exact strict path="/bridge" component={Bridge} />
                   <Route exact strict path="/farm" component={Farm} />
-                  <Route exact path="/farm/:stakingContractAddress" component={FarmSelect} />
+                  <Route exact path="/farm/:stakingContractAddress" component={FarmReward} />
                   <Route exact strict path="/lending" component={Lending} />
                   <Route component={RedirectToDefault} />
                 </Switch>
